movie-database-project-backend/pyproject.toml
[project]
name = "movie-database-project-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "cryptography>=46.0.3",
    "fastapi>=0.120.4",
    "httpx>=0.28.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.38.0",
    "yappi>=1.7.3",
]
